feat: implement caching for grades and update repository methods for context support

+++ ai-native-sdlc/server/grades-service/internal/models/models.go
@@ -4,5 +4,5 @@ type Grade struct {
   ID        int    `json:"id"`
   StudentID int    `json:"student_id" validate:"required"`
   Course    string `json:"course" validate:"required"`
-  Score     int    `json:"score" validate:"min=0,max=100"`
+  Score     float64 `json:"score" validate:"min=0,max=100"`
 }
